KOGITO-5533: add AbstractQueryEntrypointGenerator to reduce duplication
